+++ packages/libs/react-ui/src/styles/themeUtils.ts
@@ -1,3 +1,4 @@
+/* eslint @typescript-eslint/no-explicit-any: 0 */
 import type { StyleRule } from '@vanilla-extract/css';
 import type { Properties } from 'csstype';
 import omit from 'lodash.omit';
@@ -5,11 +6,11 @@ import omit from 'lodash.omit';
 // eslint-disable-next-line @kadena-dev/typedef-var
 export const breakpoints = {
   xs: '',
-  sm: `screen and (min-width: ${640 / 16}rem)`,
-  md: `screen and (min-width: ${768 / 16}rem)`,
-  lg: `screen and (min-width: ${1024 / 16}rem)`,
-  xl: `screen and (min-width: ${1280 / 16}rem)`,
-  xxl: `screen and (min-width: ${1536 / 16}rem)`,
+  sm: 'screen and (min-width: 40rem)',
+  md: 'screen and (min-width: 48rem)',
+  lg: 'screen and (min-width: 64rem)',
+  xl: 'screen and (min-width: 80rem)',
+  xxl: 'screen and (min-width: 96rem)',
 };
 
 export type Breakpoint = keyof typeof breakpoints;
@@ -71,3 +72,55 @@ export const mapToProperty =
       ? responsiveStyle({ [breakpoint]: styleRule })
       : styleRule;
   };
+
+type Token = string | { [key: string]: Token };
+type IgnoredToken = '@hover' | '@focus' | '@disabled';
+
+// eslint-disable-next-line
+const ignoredTokens = ['@hover', '@focus', '@disabled'];
+
+function isValue(token: Token): token is string {
+  return typeof token === 'string';
+}
+
+type Leaves<T> = T extends object
+  ? {
+      [K in keyof T]: `${Exclude<K, symbol | IgnoredToken>}${T[K] extends object
+        ? `.${Leaves<T[K]>}`
+        : ''}`;
+    }[keyof T]
+  : never;
+
+type FlattenObjectTokens<T extends { [key: string]: Token }> = {
+  [Key in Leaves<T>]: string;
+};
+
+/**
+ * @private Used internally to create utility class options
+ * @param {Record<string, any>} tokens - The tokens to flatten
+ * @param {string | undefined} prefix - Do not use this parameter. This param is used to internally recursively pass parent prefixes to nested tokens.
+ */
+export const flattenTokens = <T extends Record<string, any>>(
+  tokens: T,
+  prefix?: string,
+): FlattenObjectTokens<T> => {
+  if (isValue(tokens)) {
+    return { [prefix!]: tokens } as any;
+  }
+
+  const flattenedTokens: any = {};
+  Object.keys(tokens).forEach((key) => {
+    if (ignoredTokens.includes(key)) {
+      return;
+    }
+
+    const newKey = prefix !== undefined ? prefix.concat('.', key) : key;
+    const item = tokens[key];
+    if (isValue(item)) {
+      flattenedTokens[newKey] = item;
+    } else {
+      Object.assign(flattenedTokens, flattenTokens(item, newKey));
+    }
+  });
+  return flattenedTokens;
+};